Performed major restructuring and achieved basic layout of top level components
This commit makes some major changes to the file structure and react component hierarchy by removing a lot of the 'Content' and 'View' level components (see #18). This commit also marks some pre-existing components as deprecated (as well as their corresponding css files). Some files were deprecated and placed into deprecated-components instead of being entirely removed since they contain code that may be useful in the reimplementation of the top level components (FriendsView, GroupsView, SearchView).

Another important note is the the Login component now redirects the user to /chat after login credentials succesfully authenticates the user. This was done since the HomeView/HomeViewContent components were removed A TODO has been left regarding the name of the path.

The basic structure of the 3 top level components have also been achieved by this commit. This required some changes to be made to various css files and the renaming of selectors for better semantics.

Note that import paths in some of the components (especially for components in the deprecated folder) may be incorrect. This will be resolved in a future commit.

Currently, there is an issue with comm-entity-actions.js (there is a component that depends on this file). This will be resolved in a separate commit.l
