+++ src/client/components/ChatPane.jsx
@@ -1,5 +1,6 @@
 import React, {useState, useEffect} from "react"
-import "./views-sub-components-css-files/chat-pane-styles.css";
+import "../css-files/view-styles.css";
+import "../css-files/chat-pane-styles.css";
 export default function ChatPane(props){
 
     useEffect(() => {
@@ -14,7 +15,7 @@ export default function ChatPane(props){
     }, []);
 
     return(
-        <div id="chatPaneWrapper">
+        <div className="chatPane">
             <div id="chatBar"></div>
             <div id="chatView">
                 <div id="messagesDiv">
