+++ src/client/components/GroupsView.jsx
