+++ src/client/components/SearchBar.jsx
@@ -1,8 +1,8 @@
 import React from "react";
-import SEARCH_EVENTS from "../../../../events/search-events";
-import socket from "../../../../index";
+import SEARCH_EVENTS from "../../events/search-events";
+import socket from "../../index";
 import "./views-sub-components-css-files/search-bar-styles.css";
-import NotificationHandler from "../../../notification-handler";
+import NotificationHandler from "../notification-handler";
 
 export default class SearchBar extends React.Component{
     componentDidMount(){
