+++ src/client/components/deprecated-components/UserInfoForm.jsx
@@ -1,7 +1,9 @@
+// ################# DEPRECATED (SEE ISSUE #18) FOR DETAILS ################# //
+
 import React, {useState, useEffect} from "react"
-import socket from "../../../../index";
-import PROFILE_EVENTS from "../../../../events/profile-events";
-import NotificationHandler from "../../../notification-handler";
+import socket from "../../../index";
+import PROFILE_EVENTS from "../../../events/profile-events";
+import NotificationHandler from "../../notification-handler";
 import "./views-sub-components-css-files/user-info-form-styles.css";
 export default function UserInfoForm(props){
     const thisUser = JSON.parse(sessionStorage.getItem("thisUser"));
