+++ src/client/components/deprecated-components/profile-view-styles.css
@@ -1,3 +1,5 @@
+/* ################# DEPRECATED (SEE ISSUE #18) FOR DETAILS ################# */
+
 #profileViewWrapper{
     width: 100%;
     height: 100%;
