--- src/client/components/views-content-components/views-content-css-files/communication-view-styles.css
@@ -1,19 +1,0 @@
-#communicationViewWrapper{
-    display: flex;
-    width: 100%;
-    height: 100%;
-}
-
-#communicationViewWrapper > div {
-    margin: 1%;
-}
-
-#communicationViewSideBarWrapper{
-    display: flex;
-    flex-direction: column;
-    flex: 1;
-}
-
-#communicationEntitySearchBarWrapper{
-    flex: 1;
-}
