--- src/client/components/views-content-components/views-sub-components/SearchResultsBar.jsx
@@ -1,10 +1,0 @@
-import React, {useState, useEffect} from "react";
-import "./views-sub-components-css-files/search-results-bar-styles.css";
-import CommunicationEntity from "./CommunicationEntity";
-import COMM_ENTITY_ACTIONS from "../../../side-bar-sub-components/comm-entity-actions";
-import socket from "../../../../../index";
-import USER_ACTION_EVENTS from "../../../../../events/user-action-events";
-import NotificationHandler from "../../../../notification-handler";
-export default function SearchResultsBar(props){
-
-}
