--- src/client/components/views-content-components/views-sub-components/comm-entity-actions.js
@@ -1,9 +1,0 @@
-const COMM_ENTITY_ACTIONS = {
-    ADD: "Add",
-    REMOVE: "Remove",
-    JOIN: "Join",
-    LEAVE: "Leave",
-    DISMISS: "Dismiss"
-}
-
-export default COMM_ENTITY_ACTIONS;
