--- src/client/components/views-content-components/views-sub-components/views-sub-components-css-files/search-results-bar-styles.css
@@ -1,38 +1,0 @@
-#searchResultsBar{
-    width: 95%;
-    height: 83.3%;
-    margin: auto;
-    margin-bottom: 2%;
-    display: flex;
-    flex-direction: column;
-}
-
-.searchResultCommEntity{
-    height: 20%;
-}
-
-#searchResultsBar > div > .communicationEntityActionsDiv{
-    height: 20%;
-}
-
-#searchResultsBar > div{
-    flex: 1;
-}
-
-#searchResultsBar, #searchResultsBar > div{
-    overflow-y: auto;
-    overflow-x: hidden;
-}
-
-.resultsTitleDiv{
-    background-color: var(--accent-color);
-    color: var(--primary-color);
-    text-align: left;
-    width: 100%;
-    height: 5%
-}
-
-.resultsTitleDiv > h3{
-    font-size: 0.75vmax;
-    margin-left: 3%;
-}
