--- src/client/css-files/chat-divs.css
@@ -1,20 +1,0 @@
-#mainWrapper{
-    display: flex;
-    flex-direction: row;
-    width: 100%;
-    height: 100vh;
-}
-
-#sideBar{
-    flex: 1;
-    /* display: flex; */
-    /* flex-direction: column; */
-    background-color: var(--secondary-color);
-    /* border-radius: 0 15px 15px 0; */
-    color: var(--light-text-color);
-}
-
-#mainView{
-    flex: 5;
-    background-color: var(--primary-color);
-}
