+++ src/client/css-files/chat-pane-styles.css
@@ -1,10 +1,3 @@
-#chatPaneWrapper{
-    flex: 3;
-    border: 1px solid var(--grey-border-color);
-    display: flex;
-    flex-direction: column;
-}
-
 /* chatTitle and chatView flex values based of ratios between SearchBar and CommunicationEntityBar flex values*/
 #chatBar{
     flex: 1;
@@ -62,3 +55,4 @@
     transition: 0.2s linear;
     border-radius: 20px;
 }
+
