+++ src/client/css-files/create-group-view-styles.css
