+++ src/client/css-files/search-view-styles.css
